Move the UI support for device protection to pyanaconda.ui.lib
Remove the storage argument from the functions mark_protected_device and
unmark_protected_device and move them to pyanaconda.ui.lib.
