+++ pyanaconda/ui/helpers.py
@@ -56,7 +56,7 @@
 from abc import ABCMeta, abstractproperty, abstractmethod
 
 from pyanaconda.core import constants
-from pyanaconda.storage.utils import unmark_protected_device, mark_protected_device
+from pyanaconda.ui.lib.storage import mark_protected_device, unmark_protected_device
 from pyanaconda.threading import threadMgr
 from pyanaconda.payload.manager import payloadMgr
 from pyanaconda.anaconda_loggers import get_module_logger
@@ -121,7 +121,7 @@ class SourceSwitchHandler(object, metaclass=ABCMeta):
         so it can be used for the installation.
         """
         if self.data.method.method == "harddrive" and self.data.method.partition:
-            unmark_protected_device(self.storage, self.data.method.partition)
+            unmark_protected_device(self.data.method.partition)
 
     def set_source_hdd_iso(self, device_name, iso_path):
         """ Switch to the HDD ISO install source
@@ -140,7 +140,7 @@ class SourceSwitchHandler(object, metaclass=ABCMeta):
             self._clean_hdd_iso()
 
         # protect current device_name
-        mark_protected_device(self.storage, device_name)
+        mark_protected_device(device_name)
 
         self.data.method.method = "harddrive"
         self.data.method.partition = partition
